refactor: centralize protobuf code generation
