+++ codegen/regen.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+set -e
+
+repo_root=$(git rev-parse --show-toplevel)
+cd "$repo_root" && cargo run --manifest-path codegen/Cargo.toml
