+++ src/protocol/grpc/observability/README.md
@@ -5,8 +5,9 @@ This module provides gRPC-based observability services using Protocol Buffers.
 The setup is inspired by [Apache DataFusion's proto crate](https://github.com/apache/datafusion/tree/main/datafusion/proto) structure but adapted
 for tonic/gRPC services.
 
-In the root of the datafusion-distribued repo, run:
+The centralized generator rebuilds all protobuf stubs, including this module.
+In the root of the datafusion-distributed repo, run:
 
 ```bash
-./src/protocol/grpc/observability/gen/regen.sh
+./codegen/regen.sh
 ```
