--- src/protocol/grpc/observability/gen/regen.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-set -e
-
-repo_root=$(git rev-parse --show-toplevel)
-cd "$repo_root" && cargo run --manifest-path src/protocol/grpc/observability/gen/Cargo.toml
